# Refactor: wtype per tensor from file instead of global