--- a/control.hpp
+++ b/control.hpp
@@ -317,10 +317,12 @@ struct ControlNet : public GGMLRunner {
     bool guided_hint_cached         = false;
 
     ControlNet(ggml_backend_t backend,
-               ggml_type wtype,
                SDVersion version = VERSION_SD1)
-        : GGMLRunner(backend, wtype), control_net(version) {
-        control_net.init(params_ctx, wtype);
+        : GGMLRunner(backend), control_net(version) {
+    }
+
+    void init_params(std::map<std::string, enum ggml_type>& tensor_types, const std::string prefix) {
+        control_net.init(params_ctx, tensor_types, prefix);
     }
 
     ~ControlNet() {