--- a/esrgan.hpp
+++ b/esrgan.hpp
@@ -142,12 +142,11 @@ struct ESRGAN : public GGMLRunner {
     int scale     = 4;
     int tile_size = 128;  // avoid cuda OOM for 4gb VRAM
 
-    ESRGAN(ggml_backend_t backend)
+    ESRGAN(ggml_backend_t backend,std::map<std::string, enum ggml_type>& tensor_types)
         : GGMLRunner(backend) {
+        rrdb_net.init(params_ctx, tensor_types, "");
     }
-    void init_params(std::map<std::string, enum ggml_type>& tensor_types, const std::string prefix) {
-        rrdb_net.init(params_ctx, tensor_types, prefix);
-    }
+
 
     std::string get_desc() {
         return "esrgan";