@stduhpf 2024-11-23
Trying master again after having these changes implemented for a while it feels like this reduces the loading times, but maybe there's somthing else affecting it.