@fszontagh 2024-11-23
Maybe this is the wrong place where i write this, but i started a flux diffusion, which is loaded the model into VRAM, and started to make the diffusion on CPU. Please the the screenshot below: 

![image](https://github.com/user-attachments/assets/fed21858-cfce-4edc-a7e4-e938e8e4853e)