@Green-Sky 2024-11-24
You probably saw the embedding models.