@stduhpf 2024-11-25
I can confirm now, this PR makes loading weights very much faster for larger models.

(Results including warm runs only, so the models are always in disk cache)
| model (types) | master (model loading time) | PR (model loading time) |
| --- | --- | ---|
| flux1-schnell (q3_k, ae: q8_0, clip_l: q8_0, t5xxl: q4_k) | 92.37s | 6.08s |
| sd3.5_medium (q8_0, clip_l: f16, clip_g: f16, t5xxl: q4_k) | 18.19s | 5.37s |
| sd3.5_large (q4_0, clip_l: q8_0, clip_g: f16, t5xxl: q4_k) | 104.6 | 5.95 |
| SDXL (q8_0, vae: f16) | 2.70s | 2.72s |
| SD 1.5 (f16) | 1.42s | 1.42s |

I think this makes this PR worth merging.