@stduhpf 2024-11-26
Ah, controlnets are not working, I'll see if I can fix.