@stduhpf 2024-11-27
I think I got pretty much everything working at least as well as it used to with this refactoring now.  If anyone notices something I might have missed, lmk.