@thxCode 2024-11-29
it introduces a tensor types mapper to guide the ggml type of each tensor, but i saw the conversion has filtered out these fixed tensors, is this an overkill implementation? https://github.com/leejet/stable-diffusion.cpp/blob/master/model.cpp#L1873-L1902

the root cause is using one `Conditioner weight type` to replace the weight types of various clip_l/clip_g/t5xxl, why do we still keep this wrong information `Conditioner weight type:     fp16`(clip_l: fp16, clip_g: fp16, t5xxl: q8_0)?

can we just refactor the `get_conditioner_wtype` of the model loader, make it able to extract the ggml type of each conditioner according to the sd_version, and then use the correct ggml type to initialize the different conditioners? looks like changing is less but achieving the same goals.