@stduhpf 2024-11-29
> can we just refactor the get_conditioner_wtype of the model loader, make it able to extract the ggml type of each conditioner according to the sd_version, and then use the correct ggml type to initialize the different conditioners? looks like changing is less but achieving the same goals.

That's a fair point. It would indeed probably improve the loading times the same way without refactoring the whole thing. But the point of this PR was initially to refactor the model loading logic, the improvement in loading time for conditionning models is just a nice side effect.

My original motivation for doing this refactor was to be able to better support models with mixed quantization types (like those made with https://github.com/city96/ComfyUI-GGUF/blob/main/tools/convert.py or with https://github.com/leejet/stable-diffusion.cpp/pull/447). Now it also make it possible to implement https://github.com/leejet/stable-diffusion.cpp/pull/490 using the keys of the same tensor types map.